don't delete companion when snapshot deleted; be resilient during create if companion already exists

+++ core/src/main/scala/org/broadinstitute/dsde/rawls/snapshot/SnapshotService.scala
@@ -43,31 +43,36 @@ class SnapshotService(protected val userInfo: UserInfo, val dataSource: SlickDat
 
       val referenceId = snapshotRef.getMetadata.getResourceId
 
-      // create BQ dataset
-      val createDatasetIO = deltaLayer.createDataset(workspaceContext, userInfo)
+      // attempt to create the BQ dataset, which might already exist
+      val createDatasetIO = deltaLayer.createDatasetIfNotExist(workspaceContext, userInfo)
 
       createDatasetIO.unsafeToFuture().recover {
         case t: Throwable =>
           //fire and forget this undo, we've made our best effort to fix things at this point
           IO.pure(workspaceManagerDAO.deleteDataRepoSnapshotReference(workspaceContext.workspaceIdAsUUID, referenceId, userInfo.accessToken)).unsafeToFuture()
           throw new RawlsExceptionWithErrorReport(ErrorReport(StatusCodes.InternalServerError, s"Unable to create snapshot reference in workspace ${workspaceContext.workspaceId}. Error: ${t.getMessage}"))
-      }.flatMap { datasetId =>
+      }.flatMap {
+        case (justCreated, datasetId) =>
+          // if we just created the companion dataset - because it didn't exist - then also create a WSM reference to it
+          if (justCreated) {
+            val createBQReferenceFuture = for {
+              petToken <- samDAO.getPetServiceAccountToken(GoogleProjectId(workspaceName.namespace), SamDAO.defaultScopes + SamDAO.bigQueryReadOnlyScope, userInfo)
+              bigQueryRef = workspaceManagerDAO.createBigQueryDatasetReference(workspaceContext.workspaceIdAsUUID, new ReferenceResourceCommonFields().name(datasetId.getDataset).cloningInstructions(CloningInstructionsEnum.NOTHING), new GcpBigQueryDatasetAttributes().projectId(workspaceContext.namespace).datasetId(datasetId.getDataset), OAuth2BearerToken(petToken))
+            } yield { bigQueryRef }
 
-        // TODO: AS-724 - don't create a new ref to the companion dataset every time we add a snapshot
-        val createBQReferenceFuture = for {
-          petToken <- samDAO.getPetServiceAccountToken(GoogleProjectId(workspaceName.namespace), SamDAO.defaultScopes + SamDAO.bigQueryReadOnlyScope, userInfo)
-          bigQueryRef = workspaceManagerDAO.createBigQueryDatasetReference(workspaceContext.workspaceIdAsUUID, new ReferenceResourceCommonFields().name(datasetId.getDataset).cloningInstructions(CloningInstructionsEnum.NOTHING), new GcpBigQueryDatasetAttributes().projectId(workspaceContext.namespace).datasetId(datasetId.getDataset), OAuth2BearerToken(petToken))
-        } yield { bigQueryRef }
-
-        createBQReferenceFuture.recover {
-          case t: Throwable =>
-            //fire and forget these undos, we've made our best effort to fix things at this point
-            for {
-              _ <- deltaLayer.deleteDataset(workspaceContext).unsafeToFuture()
-              _ <- Future(workspaceManagerDAO.deleteDataRepoSnapshotReference(workspaceContext.workspaceIdAsUUID, referenceId, userInfo.accessToken))
-            } yield {}
-            throw new RawlsExceptionWithErrorReport(ErrorReport(StatusCodes.InternalServerError, s"Unable to create snapshot reference in workspace ${workspaceContext.workspaceId}. Error: ${t.getMessage}"))
-        }
+            createBQReferenceFuture.recover {
+              case t: Throwable =>
+                //fire and forget these undos, we've made our best effort to fix things at this point
+                for {
+                  _ <- deltaLayer.deleteDataset(workspaceContext).unsafeToFuture()
+                  _ <- Future(workspaceManagerDAO.deleteDataRepoSnapshotReference(workspaceContext.workspaceIdAsUUID, referenceId, userInfo.accessToken))
+                } yield {}
+                throw new RawlsExceptionWithErrorReport(ErrorReport(StatusCodes.InternalServerError, s"Unable to create snapshot reference in workspace ${workspaceContext.workspaceId}. Error: ${t.getMessage}"))
+            }
+          } else {
+            Future(workspaceManagerDAO.getBigQueryDatasetReferenceByName(workspaceContext.workspaceIdAsUUID, datasetId.getDataset, userInfo.accessToken))
+            // retrieve existing dataset reference
+          }
       }.map { _ => snapshotRef}
     }
   }
@@ -113,16 +118,6 @@ class SnapshotService(protected val userInfo: UserInfo, val dataSource: SlickDat
       // check that snapshot exists before deleting it. If the snapshot does not exist, the GET attempt will throw a 404
       val snapshotRef = workspaceManagerDAO.getDataRepoSnapshotReference(workspaceContext.workspaceIdAsUUID, snapshotUuid, userInfo.accessToken)
       workspaceManagerDAO.deleteDataRepoSnapshotReference(workspaceContext.workspaceIdAsUUID, snapshotUuid, userInfo.accessToken)
-
-      val datasetName = DeltaLayer.generateDatasetNameForReference(snapshotRef.getMetadata.getResourceId)
-      deltaLayer.deleteDataset(workspaceContext).unsafeToFuture().map { _ =>
-        val datasetRef = workspaceManagerDAO.getBigQueryDatasetReferenceByName(workspaceContext.workspaceIdAsUUID, datasetName, userInfo.accessToken)
-        workspaceManagerDAO.deleteBigQueryDatasetReference(workspaceContext.workspaceIdAsUUID, datasetRef.getMetadata.getResourceId, userInfo.accessToken)
-      }.recover {
-        case t: Throwable =>
-          logger.warn(s"A snapshot reference was deleted, but an error occurred while deleting its Delta Layer companion dataset: snapshot ref ID: ${snapshotRef.getMetadata.getResourceId}, dataset name: ${datasetName}")
-          throw new RawlsExceptionWithErrorReport(ErrorReport(StatusCodes.InternalServerError, s"Your snapshot reference was deleted, but an error occurred while deleting its Delta Layer companion dataset. Error: ${t.getMessage}"))
-      }
     }
   }
